fix for tcllib websocket (un)masking
The websocket package from tcllib seems to have a problem with
unmasking. It was done using unsigned ints read as big endian (binary
scan Iu) and on whole words in websocket::Mask, but this hangs in some
cases, encountered for binary data.
Changed the (un)masking to read the mask a list of 4 bytes from the ws
packet header and do bytewise masking of the data. This is probably
slower, but it works.